Return empty set when searching for non-existent name
Currently, searching for a name that doesn't exist returns not the empty set,
but the set with one member, nil.

This is because searching by name works by looking up the name in a hash, and
adding the result of that lookup to the result set. When the name doesn't
exist, the hash lookup fails and returns nil, which gets added to the result
set.

Fix this by checking whether the hash lookup returned nil, and if so, don't do
anything since the result set is already empty.

Other kinds of search do not have the same bug, since they work by filtering
lists, not blindly adding elements to the result set.