Simple script for writing a partial Kea configuration
Another simple script, mostly as an example of how to use Palletjack. This one
writes a JSON fragment to stdout, containing subnet definitions for a Kea DHCP
server with static reservations as defined in the warehouse.

It does not yet write an entire Kea configuration, since it is missing a few
global directives.