Split project into two gems, one for the library and one for the tools
Create a new gem, palletjack-tools, for the binaries in tools/. Make it
depend on exactly the same version of palletjack for now, since API:s are in
flux this early in development. Consider changing it to a pessimistic
version dependency once things calm down.

This also necessitates bumping both version numbers to 0.0.3, since the tools
now depend on functionality introduced to the library after 0.0.2.