First draft of example warehouse
We need an example warehouse in the source distribution to show how this is
supposed to work. This is a first draft of changing the private example to a
public one. It's been run through palletjack2pxelinux and palletjack2kea, so
it is at least somewhat internally consistent. It is designed for upstream
master, i.e. it does not take pull request #9 into consideration.

Please check that there is no private information left, and that it actually
works as designed.