First, untested, attempt at palletjack2salt
A program that takes a Salt minion ID and outputs a YAML structure with the
data Salt needs for the system with that ID. Meant to be called as a cmd_yaml
external pillar, e.g. by this Salt master config:

ext_pillar:
- cmd_yaml: '/usr/bin/palletjack2salt --warehouse /etc/warehouse/ --minion %s'

Not tested yet, because I have no machine with both a Salt test environment
and a Pallet Jack development environment.