Don't write reservations for IPv4 interfaces without MAC addresses
The example warehouse contains dangling IPv4 interfaces with just an IPv4
address and a hostname, so that palletjack2knot will generate A records for
them in DNS. This confuses palletjack2kea, which generates invalid DHCP
reservations for those addresses.

Make palletjack2kea skip interfaces without a MAC address to avoid this
problem.