Omit the commit ID of HEAD by default
Since we want to re-run all the palletjack2* tools on each commit to the
warehouse repository, we only want their output to change if the actual data
changes. If the header includes the commit ID, it will always change, and Salt
will always update all files and restart all services.