When reading the warehouse, ignore hidden files
Emacs uses lock files in the form of symbolic links called .#<filename> and
pointing to a string containing username, hostname and PID of the Emacs
process currently editing a file. Pallet Jack gets confused when reading
these, so they need to be skipped. Skipping everything starting with . has
the additional advantage of not surprising the user with hidden data.