Make pallets remember their own types and names
Currently, the name of a pallet is stored in the key `pallet[kind]`. However,
the kind is not stored anywhere. This makes it difficult to find out anything
about a pallet object without context.

Make each pallet remember its own type and name in the instance variables
`kind` and `name`, respectively.