Use keyvaluedag child relations to find IPv4 interfaces in an IPv4 network
Currently, palletjack2kea uses the Pallet Jack search functions to find the
interfaces belonging to each network. This makes a linear scan over the entire
warehouse object graph, which is needlessly wasteful. Instead, use the fact
that we know that all interfaces will be direct children of the network, and
traverse the child list, picking out all objects of the correct type.