Name keys in the `pallet.references` namespace according to source, not target
Currently, when a warehouse object contains a symlink, the link target is
added to the `pallet.references` namespace with the key set according to the
object's type. Since we want to be able to distinguish between different kinds
of inheritance, change this to instead take the key from the name of the link
source. This means you can have e.g.

   ipv4_interface/X.X.X.X/virtual_nic -> ../../phy_nic/Y:Y:Y:Y:Y:Y

and get the keys

   pallet.references.virtual_nic: Y:Y:Y:Y:Y:Y
   pallet.phy_nic: Y:Y:Y:Y:Y:Y

The `pallet` namespace will still contain keys according to the target type,
since they are set by the target objects themselves, which do not know how
they are being pointed to.