Clarify rdoc and API change for #pallet_link
Disallow the short form `link_type:[...]` for declaring a parent link,
where `link_type` was explicitly converted to a string.

Instead use the long form `link_type => [...]`, where `link_type`
will be looked up in `options` if it is a symbol.

This unifies the semantics of symbol lookup via `#config_path` for
all filesystem related components. Updated documentation and tool
implementations to match this change.