Define API for testable split tool functionality
To easily test the tools, we want them to be able to store their data
in an intermediate format, easily compared, instead of having the test
framework read text files from disk.

Add the functions #process and #output to the Tool class for this
purpose. Override these functions in specific tool classes.