Refactor pallet creation
Refactor the (internal) API for creating and loading pallets.

Move the responsibility for managing the pallet lookup cache from pallets to the pallet jack. This adds a new public API method for loading a named pallet
directly, `PalletJack#pallet(KIND, NAME)`, which is also used to create and load new pallets internally. Also, `PalletJack#warehouse` gives public access to the base path of the warehouse directory.