Refactor PalletJack namespace
Having the top level name space for palletjack depending on `KVDAG`
turned out to cause all sorts of problems with bundler.

Refactor all functionality into `PalletJack::Warehouse` instead, and
change the top level `PalletJack` to be a `Module` instead of a `Class`.

Since the public class API for PalletJack only had `.load`, and calling
`.new` was deprecated, keep `.load` as a convenience method for loading a
`PalletJack::Warehouse`.